+++ java/client/src/main/java/org/signal/libsignal/net/UnauthenticatedChatConnection.java
@@ -33,7 +33,7 @@ public class UnauthenticatedChatConnection extends ChatConnection {
     this.keyTransparencyClient = new KeyTransparencyClient(this, tokioAsyncContext, ktEnvironment);
   }
 
-  private KeyTransparencyClient keyTransparencyClient;
+  private final KeyTransparencyClient keyTransparencyClient;
 
   static CompletableFuture<UnauthenticatedChatConnection> connect(
       final TokioAsyncContext tokioAsyncContext,