+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 v0.92.2
 
+
+- keytrans: Periodically update distinguished tree head inside libsignal